[#47] LocationManager 분리 중간과정
