+++ RunLog/RunLog/Sources/Presentation/Run/View/RunHomeViewController.swift
@@ -112,7 +112,7 @@ final class RunHomeViewController: UIViewController {
         startButton.publisher
             .sink {
                 print("운동 시작하기 버튼 클릭")
-                LocationManager.shared.isRunning = true
+//                LocationManager.shared.isRunning = true
                 PedometerManager.shared.startPedometerUpdate()
                 let vc = RunningViewController()
                 vc.modalPresentationStyle = .fullScreen
